refactor: split lifecycle into separate crate
